Use FEATURE_* if/def statements to include platform specific features
Some features were behind if/def statements, but both DESKTOP and the name of a feature was used for those statements. None of them were enabled in the proj files, so they never were enabled in any build. This replaces those if/def statements with names that explicitly say what the feature that is needed for the code to be enabled.