Abstract input assemblies into IAssemblyFile
Using a FileInfo required that every input be a file. However, there are
circumstances where this is not desired, such as reading from a stream
or in a webservice. This change abstracts the necessary file information
for an input assembly that can then be implemented for a concrete
assembly input.