Add links to Channel 9 video
We recently did an interview for Channel 9 that does a good job of
providing an overview of API Port and explains generally how the
tool should be used and how it works. It also contains a demo.

Unfortunately, GitHub doesn't support embedding videos, so I've
cheated by referencing the image and adding a link that allows
clicking it. The video must be watched outside of GitHub.

I've added the links in two places:

* The README.md file. This is important as it serves as the landing
  page on GitHub.

* The documentation. Given that several places will link to it
  directly, it also makes sense to have the video advertised there
  as well.