Set task completion for build in single location
IVsUpdateSolutionEvents.UpdateSolution_Done is called anytime a solution is done building. If the TaskCompletionSource is set in both UpdateSolution_Cancel as well as in UpdateSolution_Done, it will end up throwing (a TrySet call should be used in that case). However, the UpdateSolution_Done method receives a parameter indicating if there was a cancel, so we should just use that.