Add command 'order' to show recommended ordering of porting activities.
When looking at converting multiple projects, it works best to do a bottom-up approach so that when you start a new project, the dependencies are all available on .NET Standard/.NET Core. This provides a command to identify that.

Usage: apiport.exe -f [file or directories] -e [entrypoint]

The entry point is needed to identify what dependencies are required to get that specific app (or library) ported. The result will be a list of supplied assemblies in the recommended order.