Use IDependencyFilter when processing dependency list
This will exclude things like System.* and Microsoft.* that sometimes end up in the deployment (ie I recently saw System.Web in a report). Any thing that is analyzed is assumed to be user code and it ends up saying that any API available in the set provided (including potentially System.Web) is fair game and portable to .NET Core.