Update ProxyAuthenticationHandler.cs (#948)
Fixed a bug with the ProxyAuthenticationHandler that was not handling proxy authentication required properly.

While ProxyAuthenticationHandler has code to try to automatically deal with proxy authentication, it did not handle the case where the status code was ProxyAuthenticationRequired (which is likely the common case for that scenario). Now, the handler tests for this, and preemptively throws ProxyAuthenticationRequiredException so that the existing logic can kick in and try to authenticate.